requirements.txt
mlx>=0.31.0
mlx-lm>=0.31.0
mlx-metal>=0.31.0
transformers>=5.0.0
datasets>=4.0.0
peft>=0.18.0
trl>=1.0.0
torch>=2.4.0
lm-eval>=0.4.0
huggingface_hub>=1.0.0
accelerate>=1.0.0
sentencepiece>=0.2.0
scipy>=1.11.0
pyyaml>=6.0
numpy>=2.0.0
